A failing test in Firefox when the server returns 204
I created a new test server that was capable of returning 204s in order to pull off the test.

I also switched the "this should probably take a while" request to use another custom endpoint, since sometimes the response was actually coming back quick enough for the test to fail.