Add Option: ignoreHeadersWhenXDR
Allow caller to decide if it is an error to pass HTTP headers to CORS
requests in IE<10.

This is useful in scenarios when the library is being used to make XDR
calls and handle the difference between legacy and modern browsers
(i.e. transparently choose to use either XMLHttpRequest or XDomainRequest
on behalf of the library user).