Finnp patch 1 - Handle error events correctly (#164)
* Handle error events correctly

The `onerror` and `ontimeout` handler are not calling back with errors but with error events (without any messages). Therefore the `errorFunc` is wrapping them as `Error: [ProgressEvent object]`. 

Here's my quick proposal on how to fix this. Just pass generic error messages to the `errorFunc`, then there will be no case where the error needs to be constructed there.

* fix error messages for tests

* remove new keywords and use .bind